api: split authz runtime module
